feat(stack): make screen view component reusable
